+++ NEWS.md
@@ -1,9 +1,20 @@
+# spatialLIBD 1.1.7
+
+NEW FEATURES
+
+* The functions `sce_image_gene_p()`, `sce_image_gene()`, `sce_image_grid()`, 
+`sce_image_grid_gene()`, `sce_image_clus()`, `sce_image_clus_p()`, 
+`geom_spatial()` now work with VisiumExperiment objects thanks to the new 
+function `read_image()`. This work was done by Brenda Pardo and Leonardo. 
+
+
 # spatialLIBD 1.1.5
 
 NEW FEATURES
 
 * `fetch_data()` takes the data from sce object and creates a VisiumExperiment
-object containing these data. VisiumExperiment object can be obtained with `fetch_data("ve")`.
+object containing these data. VisiumExperiment object can be obtained with 
+`fetch_data("ve")`. This work was done by Brenda Pardo and Leonardo. 
 
 
 # spatialLIBD 1.1.4